+++ docs/conf.py
@@ -26,7 +26,7 @@ author = 'David Seddon'
 # The short X.Y version
 version = ''
 # The full version, including alpha/beta/rc tags
-release = '1.3'
+release = '2.0'
 
 
 # -- General configuration ---------------------------------------------------